app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
motor
httpx
python-jose[cryptography]
pydantic
pytest
httpx
